Fix for device detection

+++ start.sh
@@ -5,4 +5,8 @@ if [ ! -d "./workdir" ]; then
   mkdir workdir
 fi
 
+if [ -e "/dev/ttyUSB0" ]; then
 docker run -ti --rm --device /dev/ttyUSB0:/dev/ttyUSB0 -v /home/farran/dev/docker-esp-idf/workdir:/workdir espdev
+else
+docker run -ti --rm -v /home/farran/dev/docker-esp-idf/workdir:/workdir espdev
+fi
